requirements.txt
Django
gunicorn
asgiref
dj-database-url
django-cors-headers
django-heroku
djangorestframework
psycopg2
pytz
sqlparse
whitenoise
